Upd_040322 * Class String updated - fixed wrong operator '<' overload implementation. * Class EntityType updated - fixed implementation for 'RegisterScript' & 'RegisterProperty'. * Class DataType updated - fixed implementation for 'GetTypeByName'.

+++ TOD1/ScriptDatabase.cpp
@@ -122,7 +122,7 @@ int GetCommandByName(const char* const commandname)
         return GetCommandByName_Impl(commandname);
 }
 
-int RegisterGlobalProperty(const char* const propertyname, bool existingProperty)
+short RegisterGlobalProperty(const char* const propertyname, bool existingProperty)
 {
     if (!GlobalPropertiesList.size())
         GlobalPropertiesList.reserve(6100);
